# syncify.web.client/src/api/generated/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Syncify.Web.Server",
    "version": "1.0"
  },
  "paths": {
...
    "/api/recipe-ingredient": {
      "post": {
        "tags": [
          "RecipieIngredient"
        ],
        "operationId": "Create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeIngredientCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeIngredientCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeIngredientCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeIngredientGetDto>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeIngredientGetDto>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeIngredientGetDto>"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipe-ingredient/recipe/{recipeId}": {
      "get": {
        "tags": [
          "RecipieIngredient"
        ],
        "operationId": "GetAll",
        "parameters": [
          {
            "name": "recipeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<List<RecipeIngredientGetDto>>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<List<RecipeIngredientGetDto>>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<List<RecipeIngredientGetDto>>"
                }
              }
            }
          }
        }
      }
    },
    "/api/recipe-ingredient/{id}": {
      "get": {
        "tags": [
          "RecipieIngredient"
        ],
        "operationId": "GetById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeIngredientGetDto>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeIngredientGetDto>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeIngredientGetDto>"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RecipieIngredient"
        ],
        "operationId": "Update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeIngredientUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeIngredientUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeIngredientUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeIngredientGetDto>"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeIngredientGetDto>"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response<RecipeIngredientGetDto>"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "RecipieIngredient"
        ],
        "operationId": "Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmptyResponse": {
        "required": [
          "errors",
          "hasErrors"
        ],
        "type": "object",
        "properties": {
          "hasErrors": {
            "type": "boolean",
            "readOnly": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "additionalProperties": false
      },
      "Error": {
        "required": [
          "errorMessage"
        ],
        "type": "object",
        "properties": {
          "errorMessage": {
            "type": "string"
          },
          "propertyName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RecipeIngredientCreateDto": {
        "required": [
          "name",
          "quantity",
          "recipeId",
          "unit"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "unit": {
            "type": "string"
          },
          "recipeId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "RecipeIngredientGetDto": {
        "required": [
          "id",
          "name",
          "quantity",
          "recipeId",
          "unit"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "unit": {
            "type": "string"
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "recipeId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "RecipeIngredientUpdateDto": {
        "required": [
          "name",
          "quantity",
          "unit"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "unit": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "Response<List<RecipeIngredientGetDto>>": {
        "required": [
          "errors",
          "hasErrors"
        ],
        "type": "object",
        "properties": {
          "hasErrors": {
            "type": "boolean",
            "readOnly": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeIngredientGetDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Response<RecipeIngredientGetDto>": {
        "required": [
          "data",
          "errors",
          "hasErrors"
        ],
        "type": "object",
        "properties": {
          "hasErrors": {
            "type": "boolean",
            "readOnly": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            }
          },
          "data": {
            "$ref": "#/components/schemas/RecipeIngredientGetDto"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}